+++ src/components/MainStep/MainStep1/ThisMonthEmotion/ThisMonthEmotion.style.js
